* woeps * more love for the network stuff..

+++ common/PyDumpVisitor.cpp
@@ -51,7 +51,7 @@ void PyLogsysDump::_print(const char *str, ...) {
 		return;	//calling log_message directly forces us to mask it
 	va_list l;
 	va_start(l, str);
-	int len = strlen(top())+strlen(str)+2;
+	size_t len = strlen(top()) + strlen(str)+2;
 	char *buf = new char[len];
 	snprintf(buf, len, "%s%s", top(), str);
 	log_messageVA(m_type, buf, l);
@@ -77,7 +77,7 @@ PyFileDump::PyFileDump(FILE *into, bool full_hex)
 void PyFileDump::_print(const char *str, ...) {
 	va_list l;
 	va_start(l, str);
-	int len = strlen(top())+strlen(str)+2;
+	size_t len = strlen(top())+strlen(str)+2;
 	char *buf = new char[len];
 	snprintf(buf, len, "%s%s\n", top(), str);
 	vfprintf(m_into, buf, l);
@@ -311,7 +311,7 @@ void PyDumpVisitor::VisitPackedRow(const PyRepPackedRow *rep) {
 }
 
 void PyDumpVisitor::VisitString(const PyRepString *rep) {
-	if(ContainsNonPrintables(rep->value.c_str(), rep->value.length())) {
+	if(ContainsNonPrintables(rep->value.c_str(), (uint32)rep->value.length())) {
 		_print("String%s: '<binary, len=%d>'", rep->is_type_1?" (Type1)":"", rep->value.length());
 	} else {
 		_print("String%s: '%s'", rep->is_type_1?" (Type1)":"", rep->value.c_str());
